requirements.txt
pip==1.3.1
distribute==0.6.35

Django==1.6.2
django-timezones==0.2
pytz==2013d

MySQL-python==1.2.4
django-form-utils==0.3.1
South==0.8.2
django-extensions==1.3.3
django-notification==1.0
django-model-utils==1.5.0

# Fabric
Fabric==1.6.1
paramiko==1.10.1
pycrypto==2.6

# Testing dependencies
django-factory==0.11
factory-boy==2.1.2
mock==1.0.1

# django-webtest
django-webtest==1.7.7
WebTest==2.0.6
WebOb==1.2.3
six==1.3.0
waitress==0.8.5
